backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
transformers
Pillow
geopy
pymongo
ollama
requests
motor
email-validator
pytest
python-jose[cryptography]
passlib[bcrypt]
scikit-learn
joblib
numpy
cleanvision
pandas
tqdm
python-dotenv
